Add partial functionality to '/new-match' endpoint
Namely, validation of incoming player names. And, if necessary, output
of an error message. In case of passing validation, getting the Player
from the database or saving, if it is not in the database.
But for now, the creation of a new online match and redirection to its
page has not been done.